Rev 22. Changes to the script include:
Support <w:lvlOverride> elements that include both <w:lvl> and
<w:startOverride> children.

fixup_strip_toc: Don't misinterpret a blank page in front of the table
of contents as the table of contents!

fixup_title_page: Strip stray whitespace (to avoid later choking when
trying to output an <hgroup> that contains both child elements and
whitespace).