Quote the `Liberation Serif` font name
If used unquoted, this font-family name contains a keyword in CSS, and it refers to the keyword value rather than a font family with that name. If you mean to use a font with this name rather than the keyword value, you must use quotes. http://mothereff.in/font-family#0Liberation%20Serif